+++ tests/machine/x/cs/if_then_else.cs
@@ -2,9 +2,8 @@ using System;
 
 class Program {
     static void Main() {
-        long x = 12;
+        int x = 12;
         string msg = ((x > 10) ? "yes" : "no");
         Console.WriteLine(msg);
     }
 }
-
