+++ tests/machine/x/cs/if_then_else_nested.cs
@@ -2,9 +2,8 @@ using System;
 
 class Program {
     static void Main() {
-        long x = 8;
+        int x = 8;
         string msg = ((x > 10) ? "big" : ((x > 5) ? "medium" : "small"));
         Console.WriteLine(msg);
     }
 }
-
