+++ tests/machine/x/cs/len_map.cs
@@ -1,10 +1,8 @@
 using System;
 using System.Collections.Generic;
 
-class Program
-{
-    static void Main()
-    {
-        Console.WriteLine(new Dictionary<string, long> { { "a", 1 }, { "b", 2 } }.Count);
+class Program {
+    static void Main() {
+        Console.WriteLine(new Dictionary<string, int> { { "a", 1 }, { "b", 2 } }.Count);
     }
 }
