+++ tests/machine/x/cs/map_in_operator.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 class Program {
     static void Main() {
-        Dictionary<long, string> m = new Dictionary<long, string> { { 1, "a" }, { 2, "b" } };
+        Dictionary<int, string> m = new Dictionary<int, string> { { 1, "a" }, { 2, "b" } };
         Console.WriteLine(m.ContainsKey(1));
         Console.WriteLine(m.ContainsKey(3));
     }
