+++ tests/machine/x/cs/var_assignment.cs
@@ -2,7 +2,7 @@ using System;
 
 class Program {
     static void Main() {
-        long x = 1;
+        int x = 1;
         x = 2;
         Console.WriteLine(x);
     }
